# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TwinBlueprint Server API",
    "description": "REST API for TwinBlueprint — leads, bids, projects, campaigns, analytics, and admin auth.",
    "version": "3.0.0",
    "contact": {
      "name": "TwinBlueprint",
      "email": "twinblueprints@gmail.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token. Can also be sent via cookie named 'token'."
      }
    },
    "schemas": {
...
      "Pagination": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          },
          "pages": {
            "type": "integer"
          }
        }
      },
...
      "Supplier": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "role": {
            "type": "string",
            "nullable": true
          },
          "tools": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "temperature": {
            "type": "string",
            "nullable": true,
            "enum": [
              "hot",
              "warm",
              "cool"
            ]
          },
          "contact": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string",
                "nullable": true
              },
              "job_title": {
                "type": "string",
                "nullable": true
              },
              "email": {
                "type": "string",
                "format": "email",
                "nullable": true
              }
            }
          },
          "visualisation_tool": {
            "type": "string",
            "nullable": true
          },
          "uses_3d": {
            "type": "boolean",
            "nullable": true
          },
          "opportunity": {
            "type": "string",
            "nullable": true
          },
          "pain_points": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "SupplierSummary": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Supplier"
          },
          {
            "type": "object",
            "properties": {
              "active_project_count": {
                "type": "integer",
                "minimum": 0
              },
              "regions": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        ]
      },
      "SupplierIntelligence": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SupplierSummary"
          },
          {
            "type": "object",
            "properties": {
              "active_projects": {
                "type": "array",
                "items": {
                  "type": "object",
                  "required": [
                    "id",
                    "name",
                    "client",
                    "status",
                    "phase",
                    "value",
                    "currency"
                  ],
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "name": {
                      "type": "string"
                    },
                    "client": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "Active",
                        "In-flight",
                        "On Hold",
                        "Completed",
                        "Cancelled"
                      ]
                    },
                    "phase": {
                      "type": "string",
                      "enum": [
                        "Planning",
                        "Design",
                        "Construction",
                        "In Progress",
                        "Completed"
                      ]
                    },
                    "value": {
                      "type": "number",
                      "nullable": true
                    },
                    "currency": {
                      "type": "string",
                      "nullable": true
                    }
                  }
                }
              },
              "related_suppliers": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "name": {
                      "type": "string"
                    },
                    "role": {
                      "type": "string",
                      "nullable": true
                    },
                    "shared_project_count": {
                      "type": "integer",
                      "minimum": 1
                    }
                  }
                }
              },
              "opportunities": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "name": {
                      "type": "string"
                    },
                    "project": {
                      "type": "string",
                      "nullable": true
                    },
                    "value": {
                      "type": "number",
                      "nullable": true
                    },
                    "currency": {
                      "type": "string",
                      "nullable": true
                    },
                    "insight": {
                      "type": "string",
                      "nullable": true
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "open",
                        "qualified",
                        "won",
                        "lost"
                      ]
                    }
                  }
                }
              }
            }
          }
        ]
      },
...
    }
  },
  "paths": {
...
    "/api/suppliers": {
      "get": {
        "tags": [
          "Suppliers"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List supplier profiles",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Supplier intelligence",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "suppliers": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/SupplierIntelligence"
                          }
                        },
                        "pagination": {
                          "$ref": "#/components/schemas/Pagination"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          }
        },
        "description": "Includes active_project_count and active_projects for projects with lifecycle status Active or In-flight. Relationships use project_suppliers IDs when present and trimmed, case-insensitive legacy supplier names as a compatibility fallback."
      },
      "post": {
        "tags": [
          "Suppliers"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Create supplier (admin)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "role": {
                    "type": "string",
                    "nullable": true
                  },
                  "tools": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "temperature": {
                    "type": "string",
                    "nullable": true,
                    "enum": [
                      "hot",
                      "warm",
                      "cool"
                    ]
                  },
                  "contact": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string",
                        "nullable": true
                      },
                      "job_title": {
                        "type": "string",
                        "nullable": true
                      },
                      "email": {
                        "type": "string",
                        "format": "email",
                        "nullable": true
                      }
                    }
                  },
                  "visualisation_tool": {
                    "type": "string",
                    "nullable": true
                  },
                  "uses_3d": {
                    "type": "boolean",
                    "nullable": true
                  },
                  "opportunity": {
                    "type": "string",
                    "nullable": true
                  },
                  "pain_points": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Supplier profile",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "supplier": {
                          "$ref": "#/components/schemas/Supplier"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid profile"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Admin only"
          }
        }
      }
    },
    "/api/suppliers/{id}": {
      "get": {
        "tags": [
          "Suppliers"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get supplier",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Supplier intelligence",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "supplier": {
                          "$ref": "#/components/schemas/SupplierIntelligence"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "404": {
            "description": "Supplier not found"
          }
        },
        "description": "Adds active_project_count and active_projects for projects with lifecycle status Active or In-flight. Each project includes its lifecycle status, phase, nullable value, and currency. Empty relationships return active_project_count 0 and active_projects []."
      },
      "patch": {
        "tags": [
          "Suppliers"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Edit supplier (admin)",
        "description": "Omitted top-level fields are preserved. Supplied contact replaces the entire contact object. Unknown scalar fields are null; unknown lists are empty.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "minProperties": 1,
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "role": {
                    "type": "string",
                    "nullable": true
                  },
                  "tools": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "temperature": {
                    "type": "string",
                    "nullable": true,
                    "enum": [
                      "hot",
                      "warm",
                      "cool"
                    ]
                  },
                  "contact": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string",
                        "nullable": true
                      },
                      "job_title": {
                        "type": "string",
                        "nullable": true
                      },
                      "email": {
                        "type": "string",
                        "format": "email",
                        "nullable": true
                      }
                    }
                  },
                  "visualisation_tool": {
                    "type": "string",
                    "nullable": true
                  },
                  "uses_3d": {
                    "type": "boolean",
                    "nullable": true
                  },
                  "opportunity": {
                    "type": "string",
                    "nullable": true
                  },
                  "pain_points": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Supplier profile",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "supplier": {
                          "$ref": "#/components/schemas/Supplier"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid profile"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Admin only"
          },
          "404": {
            "description": "Supplier not found"
          }
        }
      }
    },
    "/api/bids/{id}/suppliers/{supplierId}": {
      "put": {
        "tags": [
          "Suppliers"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Link existing supplier to bid (admin, idempotent)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "supplierId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Link saved"
          },
          "400": {
            "description": "Invalid UUID"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Admin only"
          },
          "404": {
            "description": "Bid or supplier not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Suppliers"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Unlink supplier from bid (admin)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "supplierId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Link saved"
          },
          "400": {
            "description": "Invalid UUID"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Admin only"
          },
          "404": {
            "description": "Bid or supplier not found"
          }
        }
      }
    },
    "/api/suppliers/{supplierId}/opportunities": {
      "get": {
        "tags": [
          "Opportunities"
        ],
        "summary": "List supplier opportunities",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "supplierId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Supplier opportunities",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "opportunities": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "required": [
                              "id",
                              "name",
                              "project",
                              "value",
                              "currency",
                              "insight",
                              "status"
                            ],
                            "properties": {
                              "id": {
                                "type": "string",
                                "format": "uuid"
                              },
                              "name": {
                                "type": "string",
                                "minLength": 1
                              },
                              "project": {
                                "type": "string",
                                "nullable": true
                              },
                              "value": {
                                "type": "number",
                                "nullable": true
                              },
                              "currency": {
                                "type": "string",
                                "nullable": true
                              },
                              "insight": {
                                "type": "string",
                                "nullable": true
                              },
                              "status": {
                                "type": "string",
                                "enum": [
                                  "open",
                                  "qualified",
                                  "won",
                                  "lost"
                                ]
                              }
                            }
                          }
                        },
                        "pagination": {
                          "$ref": "#/components/schemas/Pagination"
                        }
                      }
                    }
                  }
                },
                "example": {
                  "success": true,
                  "data": {
                    "opportunities": [
                      {
                        "id": "uuid",
                        "name": "Airport expansion",
                        "project": "LAX Terminal Modernization",
                        "value": 2500000,
                        "currency": "USD",
                        "insight": "Early design influence",
                        "status": "qualified"
                      }
                    ],
                    "pagination": {
                      "page": 1,
                      "limit": 20,
                      "total": 1,
                      "pages": 1
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "404": {
            "description": "Supplier not found"
          }
        }
      }
    },
...
    "/api/opportunities/{id}/suppliers/{supplierId}": {
      "put": {
        "tags": [
          "Opportunities"
        ],
        "summary": "Link supplier to opportunity (admin)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "supplierId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Supplier link updated"
          },
          "400": {
            "description": "Invalid UUID"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Admin only"
          },
          "404": {
            "description": "Opportunity or supplier not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Opportunities"
        ],
        "summary": "Unlink supplier from opportunity (admin)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "supplierId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Supplier link updated"
          },
          "400": {
            "description": "Invalid UUID"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Admin only"
          },
          "404": {
            "description": "Opportunity or supplier not found"
          }
        }
      }
    },
    "/api/projects/{id}/suppliers/{supplierId}": {
      "put": {
        "tags": [
          "Projects"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Link supplier to project (admin, idempotent)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "supplierId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Link saved"
          },
          "400": {
            "description": "Invalid UUID"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Admin only"
          },
          "404": {
            "description": "Project or supplier not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Projects"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Unlink supplier from project (admin)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "supplierId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Link saved"
          },
          "400": {
            "description": "Invalid UUID"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Admin only"
          },
          "404": {
            "description": "Project or supplier not found"
          }
        }
      }
    },
...
  }
}